pyproject.toml
[tool.poetry]
name = "data_collection"
version = "0.1.0"
description = ""
authors = ["ICheered <tjeerd992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyAudio = "^0.2.11"
playsound = "^1.3.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
